requirements.txt
Django==2.2.6
coverage==6.4.1
django-filter==2.2.0
django-redis==4.10.0
djangorestframework==3.10.3
drf-spectacular==0.22.1
facebook-business==14.0.0
psycopg2==2.7.7
pre-commit==1.18.3
pytest-django==4.4.0
requests==2.22.0
gunicorn==20.0.4
celery==5.1.0
kafka-python==2.0.1
django-environ==0.9.0
django-crum==0.7.9
drf-extensions==0.5.0

black==21.7b0
flake8>=4.0.1
mypy>=0.942
isort>=5.10.1